Nuevas modificaciones

+++ controlador/ControladorRol.php
@@ -17,6 +17,8 @@ class ControladorRol{
 		$objetoConexion = new Conexion();
 		$objetoConexion->conectarDB();
 
+		$consulta = "SELECT * FROM empleado_rol WHERE id_empleado = $rol->id_empleado, rol_id =$rol->id_rol";
+		if(!empty($consulta))
 		$sql = "INSERT INTO empleado_rol(rol_id,empleado_id) VALUES('$rol->id_rol','$rol->id_empleado')"; 
 
 		$rol = mysqli_query($objetoConexion->con,$sql);
@@ -41,6 +43,7 @@ class ControladorRol{
 
 		// $sql = "SELECT empleado_id, rol_id FROM empleado_rol WHERE empleado_id = $id";
 		$sql = "SELECT empleado_id, rol_id FROM empleado_rol WHERE empleado_id = $id_empleado && rol_id = $id_rol";
+		// $sql = "DELETE FROM empleado_rol WHERE empleado_id = $roles->";
 		
 		$roles = mysqli_query($objetoConexion->con,$sql);
 
@@ -51,15 +54,9 @@ class ControladorRol{
 		$objetoConexion = new Conexion();
 		$objetoConexion->conectarDB();
 
-		$sql = "SELECT empleado_id, rol_id FROM empleado_rol WHERE empleado_id = $id_empleado && rol_id = $id_rol";
+		$sql = "DELETE FROM empleado_rol WHERE empleado_id = $roles->id_empleado && rol_id = $roles->id_rol";
 
-		$resultado1 = mysql_query($objetoConexion->con,$sql);
-		if(!empty($resultado1)){
-			$sql = "DELETE FROM empleado_rol WHERE empleado_id = $id_empleado && rol_id = $id_rol";
-		}else{
-			$sql = "INSERT INTO empleado_rol(rol_id,empleado_id) VALUES('$roles->id','$roles->seleccion')";
-		}
-		$resultados = mysql_query($objetoConexion->con,$sql);
+		$resultados = mysqli_query($objetoConexion->con,$sql);
 		return $resultados;
 		// $sql = "UPDATE empleado_rol SET empleado_id ='$roles->id_empleado', rol_id = '$roles->id_rol' WHERE rol_id = '$roles->id_empleado'";
